[update] rentalObjectDto 수정 및 위치 변경, rentalObjectMapper 수정
