+++ helloDaeyeo/src/main/java/com/daeyeo/helloDaeyeo/controller/RentalController.java
@@ -1,6 +1,6 @@
 package com.daeyeo.helloDaeyeo.controller;
 
-import com.daeyeo.helloDaeyeo.dto.RentalObjectDto;
+import com.daeyeo.helloDaeyeo.dto.rental.RentalObjectDto;
 import com.daeyeo.helloDaeyeo.dto.rental.RentalListDto;
 import com.daeyeo.helloDaeyeo.dto.rental.RentalRegisterDto;
 import com.daeyeo.helloDaeyeo.dto.rental.SearchSpecDto;
