+++ helloDaeyeo/src/main/java/com/daeyeo/helloDaeyeo/entity/RentalObject.java
@@ -20,9 +20,11 @@ public class RentalObject {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long objectIndex;
+
     // scId
     @ManyToOne(fetch = FetchType.LAZY)
     private SubCategory subCategory;
+
     // userId
     @ManyToOne(fetch = FetchType.LAZY)
     private Member member;
@@ -35,27 +37,34 @@ public class RentalObject {
 
     // 빌릴 대상 이름
     private String objectName;
+
     // 장소
     private String place;
+
     // 주소
     @Embedded
     private Address address;
+
     // 사용 비용
     private int usageFee;
+
     // 이용 기간
     @Embedded
     private UsagePeriod usagePeriod;
+
     // 취소 기간
     @Embedded
     private CancellationPeriod cancellationPeriod;
+
     // 최대 인원
     private int maxPerson;
+
     // 웹 사이트
     private String webSite;
+
     // 문의 전화
     @Embedded
     private Phone inquiryPhone;
-
 
     public void setSubCategory(SubCategory subCategory) {
         this.subCategory = subCategory;
